(1) included script to automatise the workflow and updated instructions to include them, (2) sanity checks in run_preComp_Integrals_MC.sh macro

+++ run_preComp_Integrals_MC.sh
@@ -19,14 +19,31 @@ cnt_hit=${4}
 seed=${5}
 
 if [ ${par} -eq 0 ]
-then cp $HOME/files/KDEeff_b${bin}_ev.root $WORKDIR
-else cp $HOME/files/KDEeff_b${bin}_od.root $WORKDIR
+then
+    if [ ! -r $HOME/files/KDEeff_b${bin}_ev.root ]; then
+	echo $HOME/files/KDEeff_b${bin}_ev.root not found
+	exit 1
+    fi
+    cp $HOME/files/KDEeff_b${bin}_ev.root $WORKDIR
+else
+    if [ ! -r $HOME/files/KDEeff_b${bin}_od.root ]; then
+	echo $HOME/files/KDEeff_b${bin}_od.root not found
+	exit 1
+    fi
+    cp $HOME/files/KDEeff_b${bin}_od.root $WORKDIR
+fi
+
+if [ ! -r $HOME/preComp_Integrals_MC ]; then
+    echo $HOME/preComp_Integrals_MC not found
+    exit 1
 fi
 cp $HOME/preComp_Integrals_MC .
 
 ./preComp_Integrals_MC ${bin} ${par} ${tag} ${cnt_hit} ${seed}
 
+if [ ! -d $HOME/tmp ]; then mkdir $HOME/tmp; fi
 cp PreIntMC_* $HOME/tmp/
+
 rm PreIntMC_*
 rm KDEeff_b*
 rm preComp_Integrals_MC
